requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
audioread==3.0.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
decorator==5.2.1
distro==1.9.0
filelock==3.17.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
narwhals==1.29.1
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai-whisper==20240930
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
protobuf==5.29.3
pyarrow==19.0.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
python-dateutil==2.9.0.post0
pytz==2025.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.23.1
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.9.0
toml==0.10.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
matplotlib
imageio[ffmpeg]
torchaudio
openai==1.65.5
librosa==0.10.2.post1
python-dotenv==1.0.1
streamlit==1.43.1
